refactor(frontend): ExerciseDetailPage を 9 ファイルに分割 (664 → 237 行)
664 行の 巨大ファイルから サブコンポーネント / QA ビュー / 整形 utility を 抽出。
重複していた header markup (主ビュー と QA ビュー で 同じ JSX) を ExerciseHeader に 統合 (DRY)。

抽出先:
- src/components/exercise/BackLink.tsx (18 行)
- src/components/exercise/ResultBadge.tsx (20 行)
- src/components/exercise/ExerciseHeader.tsx (39 行 / NEW DRY)
- src/components/exercise/ExampleBlock.tsx (47 行)
- src/components/exercise/ExecutionResultTable.tsx (73 行)
- src/components/exercise/SubmitResultPanel.tsx (59 行)
- src/components/exercise/SubmissionRow.tsx (23 行)
- src/components/exercise/QaExerciseView.tsx (147 行)
- src/utils/exerciseFormat.ts (35 行 / monacoLanguageOf / normalizeOutput / pad)

挙動 / 見た目 に 変更 なし (= 純粋 リファクタ)。

検証:
- tsc --noEmit クリーン
- eslint --max-warnings 0 クリーン
- vitest run: 1486 件 全 PASS (185 ファイル)
- npm run build 成功
